fix(tooltip): leaking TooltipComponent references
Fixes references to each of the `TooltipComponent` instances (one per opened tooltip) created by `MatTooltip` not being cleaned up. The issue comes from the fact that we're passing in the `ViewContainerRef` which ends up retaining the references until it is destroyed. This is problematic, because the tooltip could be attached to a very long-lived view container (e.g. inside the app navigation).