fix(input): label floating state not updated when using text binding inside textarea
Fixes the state that tells the form field whether the label should float not being up-to-date if the consumer is using a `textarea` with a text binding (e.g. `<textarea matInput>{{someValue}}</textarea>`). This causes a "changed after checked" error with regular change detection and the label to be hidden incorrectly with `OnPush` change detection.

Fixes #12070.